Changes to vertex import/export [WIP Export Breaks]
